+++ 1.20/Common/src/main/java/fuzs/mobplaques/client/handler/KeyBindingHandler.java
@@ -11,15 +11,19 @@ import net.minecraft.network.chat.Component;
 import net.minecraftforge.common.ForgeConfigSpec;
 
 public class KeyBindingHandler {
-    public static final KeyMapping TOGGLE_PLAQUES_KEY_MAPPING = new KeyMapping("key.togglePlaques", InputConstants.KEY_J, "key.categories.misc");
+    public static final String KEY_CATEGORY = "key.categories." + MobPlaques.MOD_ID;
+    public static final KeyMapping TOGGLE_PLAQUES_KEY_MAPPING = new KeyMapping("key.togglePlaques", InputConstants.KEY_J,
+            KEY_CATEGORY
+    );
     private static final Component ON_COMPONENT = Component.empty().append(CommonComponents.OPTION_ON).withStyle(ChatFormatting.GREEN);
     private static final Component OFF_COMPONENT = Component.empty().append(CommonComponents.OPTION_OFF).withStyle(ChatFormatting.RED);
+    public static final String KEY_MOB_PLAQUES_STATUS = "key.togglePlaques.message";
 
     public static void onClientTick$Start(Minecraft minecraft) {
         while (TOGGLE_PLAQUES_KEY_MAPPING.consumeClick()) {
             ForgeConfigSpec.ConfigValue<Boolean> enableRendering = MobPlaques.CONFIG.get(ClientConfig.class).allowRendering;
             enableRendering.set(!enableRendering.get());
-            Component component = Component.translatable("key.togglePlaques.message", enableRendering.get() ? ON_COMPONENT : OFF_COMPONENT);
+            Component component = Component.translatable(KEY_MOB_PLAQUES_STATUS, enableRendering.get() ? ON_COMPONENT : OFF_COMPONENT);
             minecraft.gui.setOverlayMessage(component, false);
         }
     }