+++ 1.20/Common/src/main/java/fuzs/mobplaques/client/helper/GuiBlitHelper.java
@@ -25,10 +25,10 @@ public class GuiBlitHelper {
 
     public static void innerBlit(PoseStack poseStack, VertexConsumer vertexConsumer, int packedLight, int minX, int maxX, int minY, int maxY, float zOffset, float u1, float u2, float v1, float v2, int color) {
         Matrix4f matrix4f = poseStack.last().pose();
-        float a = FastColor.ARGB32.alpha(color) / 255.0F;
-        float r = FastColor.ARGB32.red(color) / 255.0F;
-        float g = FastColor.ARGB32.green(color) / 255.0F;
-        float b = FastColor.ARGB32.blue(color) / 255.0F;
+        float a = FastColor.ABGR32.alpha(color) / 255.0F;
+        float r = FastColor.ABGR32.red(color) / 255.0F;
+        float g = FastColor.ABGR32.green(color) / 255.0F;
+        float b = FastColor.ABGR32.blue(color) / 255.0F;
         vertexConsumer.vertex(matrix4f, minX, minY, zOffset).color(r, g, b, a).uv(u1, v1).uv2(packedLight).endVertex();
         vertexConsumer.vertex(matrix4f, minX, maxY, zOffset).color(r, g, b, a).uv(u1, v2).uv2(packedLight).endVertex();
         vertexConsumer.vertex(matrix4f, maxX, maxY, zOffset).color(r, g, b, a).uv(u2, v2).uv2(packedLight).endVertex();
@@ -37,23 +37,22 @@ public class GuiBlitHelper {
 
     public static void fill(PoseStack poseStack, MultiBufferSource bufferSource, int packedLight, int minX, int minY, int maxX, int maxY, float zOffset, int color) {
         Matrix4f matrix4f = poseStack.last().pose();
-        int o;
         if (minX < maxX) {
-            o = minX;
+            int tmp = minX;
             minX = maxX;
-            maxX = o;
+            maxX = tmp;
         }
 
         if (minY < maxY) {
-            o = minY;
+            int tmp = minY;
             minY = maxY;
-            maxY = o;
+            maxY = tmp;
         }
 
-        float a = FastColor.ARGB32.alpha(color) / 255.0F;
-        float r = FastColor.ARGB32.red(color) / 255.0F;
-        float g = FastColor.ARGB32.green(color) / 255.0F;
-        float b = FastColor.ARGB32.blue(color) / 255.0F;
+        float a = FastColor.ABGR32.alpha(color) / 255.0F;
+        float r = FastColor.ABGR32.red(color) / 255.0F;
+        float g = FastColor.ABGR32.green(color) / 255.0F;
+        float b = FastColor.ABGR32.blue(color) / 255.0F;
         RenderType renderType = MobPlaques.CONFIG.get(ClientConfig.class).behindWalls ? RenderType.textBackgroundSeeThrough() : RenderType.textBackground();
         VertexConsumer bufferBuilder = bufferSource.getBuffer(renderType);
         bufferBuilder.vertex(matrix4f, minX, minY, zOffset).color(r, g, b, a).uv2(packedLight).endVertex();