+++ 1.20/CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
+## [v8.0.1-1.20.1] - 2024-02-14
+### Added
+- Add a config option for choosing if mob plaques should respect local lighting or always show with full brightness
+### Changed
+- Mob plaques now by default show for mobs hidden behind blocks, but only at 1/4 of the normal distance
+
 ## [v8.0.0-1.20.1] - 2023-06-27
 - Ported to Minecraft 1.20.1